validation on salary head before delete

+++ app/Http/Controllers/GradeSalaryMasterController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 use App\Models\BasicGrade;
 use App\Models\SalaryHead;
-
 use Illuminate\Http\Request;
 use App\Models\GradeWiseSalaryMaster;
+use App\Models\DeleteSalaryheadId;
+use Illuminate\Support\Facades\DB;
+
 
 
 class GradeSalaryMasterController extends Controller
@@ -24,23 +26,43 @@ class GradeSalaryMasterController extends Controller
 
         if ($method == "wid_formula") {
             $validated = $req->validate([
-                'head_title' => 'required',
+                'head_title' => 'required|unique:grade_wise_salary_masters,head_title,'.$req->id.',id,deleted_at,NULL',
                 'formulaOutput' => 'required',
                 'grade' => 'required'
             ]);
+        
             // Create a new salary head with formula
             $head_title = preg_replace('/\s+/', ' ', $req->head_title);
             $head_title = str_replace(' ', '_', $head_title); 
-           // $head_title = str_replace(' ', '_', $req->head_title); 
             $salary_head = GradeWiseSalaryMaster::create([
                 'head_title' =>  $head_title,
                 'formula' => $req->formulaOutput,
                 'method' => $req->method,
                 'grade' => $req->grade
             ]);
+         
+            //-------------------------------------------------------
+            $salary_head_id = $salary_head->id;
+            $pattern = '/\{([^}]+)\}/';
+            preg_match_all($pattern, $req->formulaOutput, $matches);
+            $keywords = $matches[1];
+
+            foreach($keywords as $val){
+             
+               $head = GradeWiseSalaryMaster::where('head_title', $val)->where('deleted_at', '=', NULL)->get();
+              if(!empty($head[0]->id)){
+                $delete_salary_head_data = array(
+                  "salary_head_id" =>   $salary_head_id,
+                  "involve_head_id" => $head[0]->id,
+                  "type" => "2"
+                );
+                $salary_head = DeleteSalaryheadId::create($delete_salary_head_data);
+              }
+            }
+            //------------------------------------------------------=
         } else {
             $validated = $req->validate([
-                'head_title' => 'required',
+                'head_title' => 'required|unique:grade_wise_salary_masters,head_title,NULL,id,deleted_at,NULL',
                 'amount' => 'required',
                 'grade' => 'required'
             ]);
@@ -116,8 +138,19 @@ class GradeSalaryMasterController extends Controller
     public function delete_basic_sal(Request $req)
     {
         $id = $req->sal_head_id;
-        $delete = GradeWiseSalaryMaster::find($id)->delete();
-        return redirect()->back()->with('message', 'Basic Grade Salary Master Deleted Successfully');
+        $check =  DeleteSalaryheadId::where('involve_head_id', $id)->where('type', '2')->get();
+        
+        if ($check->isEmpty())
+        {
+            $delete = GradeWiseSalaryMaster::find($id)->delete();
+            DeleteSalaryheadId::where('salary_head_id', $id)->where('type', '2')->delete();
+            return redirect()->back()->with('message', 'Basic Grade Salary Master Deleted Successfully');
+            
+         
+        } else {
+            return redirect()->back()->with('error', 'You Cannot delete this! Its depend on other Succces Factors');
+        }
     }
+    
 
 }
